// src/components/reception/utils/type.ts
export type PaginationModel = { page: number; pageSize: number }

export type PatientData = {
  name: string
  gender: number
  address: string
  patCccd: string
  birthday: Date
  phone: string
  famlilyName: string
  famlilyPhone: string
  monthsOld: number
  age: number
  oldPlaceTreatmentId: string
  presenterId: string
  year: number
  note: string
  email: string
  reasonExam: string
  status: boolean
  bhyt1:string
  bhyt2:string
  bhyt3:string
  bhyt4:string
  bhyt5:string
  bhyt6:string
}